Extract dependency-line parsing into its own module
Move the `# sync-with-uv` dependency-line handling out of sync_with_uv.py into
a new dependency_line.py leaf module, mirroring how repo_data.py isolates the
repo-URL mapping concern. The module holds the `_DEP_*` regexes, DepLineChange,
_normalize_package_name, and the now-public sync_dependency_line entry point;
sync_with_uv.py imports them and keeps the Changes result type plus the
process_config_text orchestrator. This trims sync_with_uv.py from ~300 to ~170
lines.

Also inline the short `_replace_span` helper at its two call sites (the rev
path and the dependency path) now that it no longer needs to be shared across
modules, and update CLAUDE.md's project structure and key-functions list.

No behaviour change; the existing tests (which all exercise process_config_text)
pass unchanged.
